Reorganize code structure by transfer types
