+++ web/packages/api/src/transfers/toEthereum/transferInterface.ts
@@ -1,0 +1,29 @@
+import { AssetRegistry } from "@snowbridge/base-types"
+import { DeliveryFee, Transfer, ValidationResult } from "../../toEthereum_v2"
+import { Context } from "../../index"
+
+export interface TransferInterface {
+    getDeliveryFee(
+        source: { sourceParaId: number; context: Context },
+        registry: AssetRegistry,
+        tokenAddress: string,
+        options?: {
+            padPercentage?: bigint
+            slippagePadPercentage?: bigint
+            defaultFee?: bigint
+            feeTokenLocation?: any
+        }
+    ): Promise<DeliveryFee>
+
+    createTransfer(
+        source: { sourceParaId: number; context: Context },
+        registry: AssetRegistry,
+        sourceAccount: string,
+        beneficiaryAccount: string,
+        tokenAddress: string,
+        amount: bigint,
+        fee: DeliveryFee
+    ): Promise<Transfer>
+
+    validateTransfer(context: Context, transfer: Transfer): Promise<ValidationResult>
+}
