+++ web/packages/api/src/xcmbuilders/toEthereum/erc20FromParachain.ts
@@ -1,0 +1,476 @@
+import { Registry } from "@polkadot/types/types"
+import { beneficiaryMultiAddress } from "../../utils"
+import {
+    bridgeLocation,
+    DOT_LOCATION,
+    erc20Location,
+    erc20LocationReanchored,
+    parachainLocation,
+    accountToLocation,
+    HERE_LOCATION,
+    isEthereumNative,
+} from "../../xcmBuilder"
+import { Asset } from "@snowbridge/base-types"
+
+export function buildParachainERC20ReceivedXcmOnDestination(
+    registry: Registry,
+    ethChainId: number,
+    tokenAddress: string,
+    transferAmount: bigint,
+    feeInDot: bigint,
+    beneficiary: string,
+    topic: string
+) {
+    let beneficiaryLocation = accountToLocation(beneficiary)
+    return registry.createType("XcmVersionedXcm", {
+        v5: [
+            {
+                reserveAssetDeposited: [
+                    {
+                        id: DOT_LOCATION,
+                        fun: {
+                            Fungible: feeInDot,
+                        },
+                    },
+                    {
+                        id: erc20Location(ethChainId, tokenAddress),
+                        fun: {
+                            Fungible: transferAmount,
+                        },
+                    },
+                ],
+            },
+            { clearOrigin: null },
+            {
+                buyExecution: {
+                    fees: {
+                        id: DOT_LOCATION,
+                        fun: {
+                            Fungible: feeInDot,
+                        },
+                    },
+                    weightLimit: "Unlimited",
+                },
+            },
+            {
+                depositAsset: {
+                    assets: {
+                        wild: {
+                            allCounted: 2,
+                        },
+                    },
+                    beneficiary: {
+                        parents: 0,
+                        interior: { x1: [beneficiaryLocation] },
+                    },
+                },
+            },
+            { setTopic: topic },
+        ],
+    })
+}
+
+function buildAssetHubXcmFromParachain(
+    ethChainId: number,
+    sourceAccount: string,
+    beneficiary: string,
+    tokenAddress: string,
+    topic: string,
+    sourceParachainId: number,
+    destinationFee: bigint,
+    feeAssetId: any
+) {
+    let {
+        hexAddress,
+        address: { kind },
+    } = beneficiaryMultiAddress(sourceAccount)
+    let sourceAccountLocation
+    switch (kind) {
+        case 1:
+            // 32 byte addresses
+            sourceAccountLocation = { accountId32: { id: hexAddress } }
+            break
+        case 2:
+            // 20 byte addresses
+            sourceAccountLocation = { accountKey20: { key: hexAddress } }
+            break
+        default:
+            throw Error(`Could not parse source address ${sourceAccount}`)
+    }
+    return [
+        // Error Handling, return everything to sender on source parachain
+        {
+            setAppendix: [
+                {
+                    depositReserveAsset: {
+                        assets: {
+                            wild: "All",
+                        },
+                        dest: { parents: 1, interior: { x1: [{ parachain: sourceParachainId }] } },
+                        xcm: [
+                            {
+                                buyExecution: {
+                                    fees: {
+                                        id: feeAssetId,
+                                        fun: {
+                                            fungible: destinationFee,
+                                        },
+                                    },
+                                    weightLimit: "Unlimited",
+                                },
+                            },
+                            {
+                                depositAsset: {
+                                    assets: {
+                                        wild: "All",
+                                    },
+                                    beneficiary: {
+                                        parents: 0,
+                                        interior: { x1: [sourceAccountLocation] },
+                                    },
+                                },
+                            },
+                            { setTopic: topic },
+                        ],
+                    },
+                },
+            ],
+        },
+        // Initiate the bridged transfer
+        {
+            initiateReserveWithdraw: {
+                assets: {
+                    Wild: {
+                        AllOf: { id: erc20Location(ethChainId, tokenAddress), fun: "Fungible" },
+                    },
+                },
+                reserve: bridgeLocation(ethChainId),
+                xcm: [
+                    {
+                        buyExecution: {
+                            fees: {
+                                id: erc20LocationReanchored(tokenAddress), // CAUTION: Must use reanchored locations.
+                                fun: {
+                                    Fungible: "1", // Offering 1 unit as fee, but it is returned to the beneficiary address.
+                                },
+                            },
+                            weight_limit: "Unlimited",
+                        },
+                    },
+                    {
+                        depositAsset: {
+                            assets: {
+                                Wild: {
+                                    AllCounted: 1,
+                                },
+                            },
+                            beneficiary: {
+                                parents: 0,
+                                interior: { x1: [{ AccountKey20: { key: beneficiary } }] },
+                            },
+                        },
+                    },
+                    {
+                        setTopic: topic,
+                    },
+                ],
+            },
+        },
+        {
+            setTopic: topic,
+        },
+    ]
+}
+
+export function buildAssetHubERC20TransferFromParachain(
+    registry: Registry,
+    ethChainId: number,
+    sourceAccount: string,
+    beneficiary: string,
+    tokenAddress: string,
+    topic: string,
+    sourceParachainId: number,
+    returnToSenderFee: bigint,
+    feeAssetId: any
+) {
+    return registry.createType("XcmVersionedXcm", {
+        v5: buildAssetHubXcmFromParachain(
+            ethChainId,
+            sourceAccount,
+            beneficiary,
+            tokenAddress,
+            topic,
+            sourceParachainId,
+            returnToSenderFee,
+            feeAssetId
+        ),
+    })
+}
+
+export function buildResultXcmAssetHubERC20TransferFromParachain(
+    registry: Registry,
+    ethChainId: number,
+    sourceAccount: string,
+    beneficiary: string,
+    tokenAddress: string,
+    topic: string,
+    transferAmount: bigint,
+    totalFee: bigint,
+    destinationFee: bigint,
+    sourceParachainId: number,
+    returnToSenderFee: bigint,
+    feeAssetId: any,
+    feeAssetIdReanchored: any
+) {
+    return registry.createType("XcmVersionedXcm", {
+        v5: [
+            {
+                withdrawAsset: [
+                    {
+                        id: feeAssetIdReanchored,
+                        fun: {
+                            Fungible: totalFee,
+                        },
+                    },
+                    {
+                        id: erc20Location(ethChainId, tokenAddress),
+                        fun: {
+                            Fungible: transferAmount,
+                        },
+                    },
+                ],
+            },
+            { clearOrigin: null },
+            {
+                buyExecution: {
+                    fees: {
+                        id: feeAssetIdReanchored,
+                        fun: {
+                            Fungible: destinationFee,
+                        },
+                    },
+                    weightLimit: "Unlimited",
+                },
+            },
+            ...buildAssetHubXcmFromParachain(
+                ethChainId,
+                sourceAccount,
+                beneficiary,
+                tokenAddress,
+                topic,
+                sourceParachainId,
+                returnToSenderFee,
+                feeAssetId
+            ),
+        ],
+    })
+}
+
+export function buildTransferXcmFromParachain(
+    registry: Registry,
+    ethChainId: number,
+    assetHubParaId: number,
+    sourceParachainId: number,
+    sourceAccount: string,
+    beneficiary: string,
+    topic: string,
+    asset: Asset,
+    tokenAmount: bigint,
+    localDOTFeeAmount: bigint,
+    totalDOTFeeAmount: bigint,
+    remoteEtherFeeAmount: bigint
+) {
+    let beneficiaryLocation = accountToLocation(beneficiary)
+    let sourceLocation = accountToLocation(sourceAccount)
+    let tokenLocation = asset.location || erc20Location(ethChainId, asset.token)
+    let assets = []
+
+    assets.push({
+        id: DOT_LOCATION,
+        fun: {
+            Fungible: totalDOTFeeAmount,
+        },
+    })
+    if (isEthereumNative(tokenLocation, ethChainId)) {
+        assets.push({
+            id: bridgeLocation(ethChainId),
+            fun: {
+                Fungible: remoteEtherFeeAmount + tokenAmount,
+            },
+        })
+    } else {
+        assets.push({
+            id: bridgeLocation(ethChainId),
+            fun: {
+                Fungible: remoteEtherFeeAmount,
+            },
+        })
+        assets.push({
+            id: tokenLocation,
+            fun: {
+                Fungible: tokenAmount,
+            },
+        })
+    }
+
+    let remoteInstructionsOnAH: any[] = [
+        {
+            setAppendix: [
+                {
+                    refundSurplus: null,
+                },
+                {
+                    depositAsset: {
+                        assets: {
+                            wild: {
+                                allCounted: 3,
+                            },
+                        },
+                        beneficiary: {
+                            parents: 0,
+                            interior: { x1: [sourceLocation] },
+                        },
+                    },
+                },
+            ],
+        },
+        {
+            initiateTransfer: {
+                destination: bridgeLocation(ethChainId),
+                remote_fees: {
+                    reserveWithdraw: {
+                        definite: [
+                            {
+                                id: bridgeLocation(ethChainId),
+                                fun: {
+                                    Fungible: remoteEtherFeeAmount,
+                                },
+                            },
+                        ],
+                    },
+                },
+                preserveOrigin: true,
+                assets: [
+                    {
+                        reserveWithdraw: {
+                            definite: [
+                                {
+                                    id: tokenLocation,
+                                    fun: {
+                                        Fungible: tokenAmount,
+                                    },
+                                },
+                            ],
+                        },
+                    },
+                ],
+                remoteXcm: [
+                    {
+                        depositAsset: {
+                            assets: {
+                                wild: {
+                                    allCounted: 3,
+                                },
+                            },
+                            beneficiary: {
+                                parents: 0,
+                                interior: { x1: [beneficiaryLocation] },
+                            },
+                        },
+                    },
+                    {
+                        setTopic: topic,
+                    },
+                ],
+            },
+        },
+        {
+            setTopic: topic,
+        },
+    ]
+    return registry.createType("XcmVersionedXcm", {
+        v5: [
+            {
+                withdrawAsset: assets,
+            },
+            {
+                payfees: {
+                    asset: {
+                        id: DOT_LOCATION,
+                        fun: {
+                            Fungible: localDOTFeeAmount,
+                        },
+                    },
+                },
+            },
+            {
+                setAppendix: [
+                    {
+                        refundSurplus: null,
+                    },
+                    {
+                        depositAsset: {
+                            assets: {
+                                wild: {
+                                    allCounted: 3,
+                                },
+                            },
+                            beneficiary: {
+                                parents: 0,
+                                interior: { x1: [sourceLocation] },
+                            },
+                        },
+                    },
+                ],
+            },
+            {
+                initiateTransfer: {
+                    destination: parachainLocation(assetHubParaId),
+                    remote_fees: {
+                        reserveWithdraw: {
+                            definite: [
+                                {
+                                    id: DOT_LOCATION,
+                                    fun: {
+                                        Fungible: totalDOTFeeAmount - localDOTFeeAmount,
+                                    },
+                                },
+                            ],
+                        },
+                    },
+                    preserveOrigin: true,
+                    assets: [
+                        {
+                            reserveWithdraw: {
+                                definite: [
+                                    {
+                                        id: bridgeLocation(ethChainId),
+                                        fun: {
+                                            Fungible: remoteEtherFeeAmount,
+                                        },
+                                    },
+                                ],
+                            },
+                        },
+                        {
+                            reserveWithdraw: {
+                                definite: [
+                                    {
+                                        id: tokenLocation,
+                                        fun: {
+                                            Fungible: tokenAmount,
+                                        },
+                                    },
+                                ],
+                            },
+                        },
+                    ],
+                    remoteXcm: remoteInstructionsOnAH,
+                },
+            },
+            {
+                setTopic: topic,
+            },
+        ],
+    })
+}
